Fix for CoderDojo/community-platform#932
Problem arose when you delete all the text from either input field on
the time picker. The time picker returns null in this scenario, which
when wrapped by moment results in "Invalid Date", which is from then on
used as the basis for any time/date updates. The only way to kick out
of this is to refresh the page.