Add a separate src of referer for CDF view
As there are already values in the database for some referers we no
longer want to capture we need a way to display them for older Dojo
leads.

Create a separate source to do that and allow all options to be valid.

Also does a bit of syntax mangling due to prettier and eslint.